Updating Coupons Controller Functions

+++ Algoriza_Project_2023BE83/Controllers/CouponsController.cs
@@ -2,7 +2,6 @@ using Microsoft.AspNetCore.Mvc;
 using Repository;
 using Service;
 using Core.Models;
-using Microsoft.AspNetCore.Authorization;
 namespace Algoriza_Project_2023BE83.Controllers;
 //[Authorize(Roles ="Admin")]
 [Route("api/[controller]")]
@@ -15,27 +14,66 @@ public class CouponsController : ControllerBase
         _couponsRepository = couponsRepository;
     }
     [HttpPost("")]
-    public IActionResult AddCoupon([FromBody] Coupons couponModel)
+    public async Task<IActionResult> AddCoupon([FromBody] Coupons couponModel)
     {
-        var couponId = _couponsRepository.AddCoupon(couponModel);
-        return Ok(couponId);
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        };
+
+        return await _couponsRepository.AddCoupon(couponModel);
     }
     [HttpPut("{id}")]
-    public async Task<IActionResult> UpdateCoupon([FromRoute]string id, [FromBody] Coupons couponModel)
+    public async Task<IActionResult> UpdateCoupon([FromBody]Coupons couponModel)
     {
-        await _couponsRepository.UpdateCoupon(id, couponModel);
-        return Ok();
+        if (couponModel == null)
+        {
+            ModelState.AddModelError("DiscountCodeCoupon", "The DiscountCodeCoupon is required.");
+        }
+        else if (couponModel.Id == default)
+        {
+            ModelState.AddModelError("Id", "The Id is required.");
+        }
+
+        else if (couponModel.Id < 0)
+        {
+            ModelState.AddModelError("Id", "The Id is Invalid. Must be greater than 0.");
+        }
+
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        };
+
+        return _couponsRepository.UpdateCoupon(couponModel);
     }
     [HttpDelete("{id}")]
-    public async Task<IActionResult> DeleteCoupon([FromRoute]string id)
+    public async Task<IActionResult> DeleteCoupon([FromForm]int id)
     {
-        await _couponsRepository.DeleteCoupon(id);
-        return Ok("Coupon Deleted Successfully");
+        if(id <= 0)
+            {
+            ModelState.AddModelError("id", "The Id is Invalid. Must be greater than 0.");
+        }
+
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        };
+
+        return _couponsRepository.DeleteCoupon(id);
     }
     [HttpPut("deactivate/{id}")]
-    public async Task<IActionResult> DeactivateCoupon([FromRoute]string id)
+    public async Task<IActionResult> DeactivateCoupon([FromForm]int id)
     {
-        await _couponsRepository.DeactivateCoupon(id);
-        return Ok();
+        if (id <= 0)
+        {
+            ModelState.AddModelError("id", "The Id is Invalid. Must be greater than 0.");
+        }
+        if (!ModelState.IsValid)
+        {
+            return BadRequest(ModelState);
+        };
+
+        return _couponsRepository.DeactivateCoupon(id);
     }
 }
